pyproject.toml
[tool.poetry]
name = "path-in-it-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

packages = [
    { include = "path_in_it/" },
    { include = "path_in_it/py.typed" },
]


[tool.poetry.dependencies]
python = ">=3.12, <3.13"
aiogram = "^3.6.0"
python-dotenv = "^1.0.0"
aiogram-dialog = "^2.1.0"
asyncpg = ">=0.29.0"
pydantic-settings = "^2.0.3"
jinja2 = "^3.1.2"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.30" }
fluentogram = "^1.1.7"
environs = "^11.0.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
alembic = "^1.13.1"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
i18n = "^0.2"


[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs/path_in_IT_bot"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
